requirements.txt
jsonpath
requests
ddddocr
pycryptodome